Crytex-154 AdminLogController Get

+++ Crytex.Model/Models/StateMachine.cs
@@ -9,7 +9,7 @@ namespace Crytex.Model.Models
     public class StateMachine : BaseEntity
     {
         public Int32 CpuLoad { get; set; }
-        public long RamLoad { get; set; }
+        public Int32 RamLoad { get; set; }
         public DateTime Date { get; set; }
         public Guid VmId { get; set; }
     }
